Change approach to configure what version extractor to use
Chances are there might be more than the two initial implemetations of
the version extractor. As current implementation with a boolean
parameter not extensible at all, use a symbolic name to configure
version extractor to be used as it can be extended easily to support
further implementations.

Valiidate and instantiate version extractor during configuration
validation.

To provide a new implementation for version extractor:

- implement com.carbonfive.db.migration.VersionExtractor
- add a new element to VersionExtractorsEnum and map the symbolic name
  to this new implementation's class

Signed-off-by: Mykola Nikishov <mn@mn.com.ua>